Fixed the alerts loading icon issue,bug id 1355628
RootCause:As a part of getting the instances stats in separate call in network monitoring pages,we changed the event from onUpdateData to onRowsChanged in singleDataSource function to get the event arguments which are used in the handler ,but onRowsChanged event is not trigerring some cases so reverted to onUpdateData.

Fix: As a work around we are binding the dataview event on the controller part and bug 1355674 is raised to incorporate the arguments for onUpdate event.
Change-Id: I4aa810f708b1687be07e38f4c80f62024fcfb4ea