Related-Bug: #1419840
Currently, on deleting a network from UI in vcenter mode, first corresponding
port group and ip-pool association is deleted on vcenter and on
successful completion,returns the status as success and UI refreshes the
grid and sometimes we are seeing the deleted network in grid.

After removing ip-pool association on vcenter, vcenter-plugin deletes
the network on API server and in the meanwhile if we query from API
server,it gives the deleted network also.

Added wait statement on web server,such that the call is returned as
success only after network is deleted from API server

Change-Id: Id6c7a2568c593be882d3d57999232dfb6aa03a6c