Adding the check in Service Instance for multiple Image name.
When there are multiple image created with same name in glance, Service Instance
was getting stoped in spawn state.
Fixed like if there are 2 image with same name and if the user select that
duplicate image, we are showing a pop up saying the error mesage and not letting
it to create the service Instance or tempalate.

Change-Id: Iadcda21e8be5cb093019cbae56cc2c475e3e19ad
Partial-Bug: #1415771