Added validation id device owner uuid is empty.
When the uuid field is empty it was throwing some error in error console.
Because of the field is empty it was not able to get the value and throw a
error. Now added the validation so that only if the uuid is present it will go
to next step or else error message is poped out saying UUID is empty.

Change-Id: If3947a72b8f7e305d2356850f5dd8cb98f96a156
Closes-Bug: #1423101