In case external gateway for l3 router remains same, but conected VN's are added or deleted then we need to manually update the route table ref to "rt_uuid_of_logical_router" in each VN.
In case when external gateway is updated or deleted, then neutron external gateway update
takes care of adding and deleting route table refs for the VN's.

Added error check in the logical router api part..

getting the virtual network using the new api
/api/tenants/config/all-virtual-networks-fields?uuid=
for scaling.

Change-Id: I856aa795b4f5389b6682a925098e724dcdc98702
Closes-Bug: #1445692

In port page when we create a port with Logical router the route table was not
getting updated fixed it.

Change-Id: I072f444424a4a3f620a83afef19812afb62e6082
Related-Bug: #1445692