Updating logical route ref in network's Route table.
In port page
In case of create and edit if the logical router is set for a port.
The logical router reference has to be added manualley in route table of the
network in the format of rt_uuidOfLogicalRouter.
Same way if the router is detached from the port the reference has to be removed
from the corresponding port's network.
Also if the logical router reference is changed the new reference has to be
updated in the network.
All the case are handled.

The api has got exported from logical router api page and reused in ports page.

Change-Id: I072f444424a4a3f620a83afef19812afb62e6082
Related-Bug: #1445692