Bug 1456920 The vxlan id was not getting shown according to configuared/Automatic.
The detail for the vxlan id configured/Autometic is fetched at the time of
reading the network UUID and stored in a variable.
When setting the values to the gried, using the stored value of vxlan mode
showing the corresponding vxlan.

If it is automatic the vxlan value  is taken from 'virtual_network_network_id'
if it is user entered vxlan value is taken from 'vxlan_network_identifier'

Change-Id: Ifbee5bea5e67946e21e6723d66bc479736c21dd2
Closes-Bug: #1456920