BugId 1459192 progress icon is shown until all data is loaded.
Issue
The icon was getting reset when the data is set in the grid because of the
loading icon was not getting shown.

Fixed:
Loading the data to the grid first and then setting the loading status.
So that the loading Icon is keep getting shown.
Once all the data is loaded we are not setting the loading message.
The Logic in Port and VN are different.
VN : until the network is fetched showing the grid once the Shared network call
is done Removing the message.
Port : showing untill all the UUID call is getting called in the last call
checking if UUIDArray length is less than or equal to 0 to remove the message.

Impliemented pages
Floating IP, IP Address Management, Logical Router, Policy, Port, Virtual
Network, Service Template, Security Group,

Change-Id: If72147b112f5e33f71ed0f87e2b96b62e8055803