Bug ID : 1462975. -> Bug : Security Group ID was not getting update at the time of edit. Issue : In JSON diff logic "configured_security_group_id" was not added as optFields because of that the values where not getting treated as difference and was not sent to api server at Edite case. Fixed : Added the value to the optFields so now it is getting modified.
-> Bug: In Edit window the default value of the Address has to be shown as 0.0.0.0/0
or ::/0 as entered.
Issue : The default value 0.0.0.0/0  was assigned to "Enter a CIDR". Because of
that whenever there is a value with 0.0.0.0/0 it was set as "Enter a CIDR"
Fixed: When new row is added i am passing the value as -1 and assigned the
default value as -1/0. So now the default value is assigned as "Enter a CIDR"
and if value is available(in edit case) the value is set to the field.

Change-Id: If66b90c661975cd49f47533f217db7845839b4b9
Closes-Bug: #1462975