Fixed some Bugs in port page.
Added validation code for the collections.
Fixed IP in Edit and adding new data is fixed.
In edit case fixed ip is disabled and subnet drop down is hidden..

ID perms id commented out in the api code for ports because of that IDperms will
not work.
Also in logical router page ID perms is commented out.

Change-Id: I307ffaef126de02d96a23a866b4ad013a0b40239
Closes-Bug: #1509938