Static Route Static route support in port page. Now the route-table is not getting created from port page insted the link is reffered to it. Related-Bug: #1528133
SRIOV Support in UI Done
Closes-Bug: #1535208

Bug Fix:
When there are no fixed ip specified it was trowing error saying
specify fixed ip fixed it.
Collection/deep validation is done.
Changed some access of [0] to getValueByJsonPath
In Create/Edit window Fat flow value is changed to port and "1 to 65535" is
added to the place holder.

Change-Id: Ib4050d824127775cdbe89676854cc236747b089a