Port page Issue
Fat flow schema change is done(was expecting an empty array).
Ordering/Formatting UX of Mirroring in port is done

The port getting created was not there in the list.
because the 2nd call was not getting sent as expected.
- Fixed it.
Closes-Bug: #1539950

If we refresh the port page and open the edit popup the SG was getting vanished
- Fixed it
Closes-Bug: #1539639

Routing policy page was not showing the collection in the popup.
- Fixed it.

Change-Id: Ibe31455a6a9f7b81e31c7291df98f221b81269ba