In Edit case of a Parent port, subinterface ref is getting set to empty - Fixed it
* In case of edit a parent port vmiRef was set to empty array because of
that the reference to the subInterface was removed - Fixed it.
* Added The name along with the UUID in the landing page grid.
* In landing page grid row It shows (N more) after 3 data exist In other place
it is followed as 2 so, Changed to 2.
* Not showing the sub-interface if it is a parent port.

vm ref in subinterface was added for DNS resolution and stats.
Now it is not required

Removed the reference of that VM from the created/edited Sub-interface.
Closes-Bug: #1540789

Removel of VM ref from the subinterface also solves the issue
Instance name is replaced by '-' under monitor-instance tab

Closes-Bug: #1538531

Change-Id: I0e198219f85904f30b2420ecac79ce332c1b29bc
Closes-Bug: #1543632