In Edit case of a Parent port, subinterface ref is getting set to empty - Fixed it
* In case of edit a parent port vmiRef was set to empty array because of
that the reference to the subInterface was removed - Fixed it.
* Added The name along with the UUID in the landing page grid.
* In landing page grid row It shows (N more) after 3 data exist In other place
it is followed as 2 so, Changed to 2.
* Not showing the sub-interface if it is a parent port.
* dropdownAutoWidth : false, is added for the drop down's to avoid the drop down
growing beyond the view area.
* Sub-interface cannot be removed In UI the checkbox is disabled
* In Delete case, If it is a parent port We show a message saying even the
subinterface will be delete do you want to conform then sending all the UUID of
children so that the children will be removed first and then the parent will be
removed.

Service Health Check is added in UI.
Closes-Bug: #1549112

API FIX
* The sub-interface remove is not supported in delete we are trying to update the
parent port so it was throwing some error - Fixed it.
* vm ref in subinterface was added for DNS resolution and stats.
Now it is not required
* The subinterface are removed and then the parentport is removed.
* The delete is handled in both delete and delete all

Removed the reference of that VM from the created/edited Sub-interface.
Closes-Bug: #1540789

The flow was like attach of VM will happen first and then the IP was created
because of that the ip was not setting to VM To avoid that
The interface IP is added first and then the compute is getting attachecd.

Removel of VM ref from the subinterface also solves the issue
Instance name is replaced by '-' under monitor-instance tab
Closes-Bug: #1538531

Port page not able to load sometimes
Issue:
TypeError: self.deviceOwner is not a function
- Placed the function return as inline.
So it will be not be looking for that variable.
Closes-Bug: #1542295

Change-Id: I0e198219f85904f30b2420ecac79ce332c1b29bc
Closes-Bug: #1543632