Grid sorting where not working - fixed.
The coloum with the formated data has to be specified with
sortable: {
    sortBy: 'formattedValue'
},

It was not specified in Port, Policy, Logical router and Routing policy.
Fixed in all the pages.

There where some unwanted files in old port folder Removed it.

Change-Id: Icabe60b525af66136104df8348e43e569e001b1f
Closes-Bug: #1544834

Closes-Bug: #1546884 - Handling the case where the cpu was available and showing
it as '-' or '--' in Monitor Infra summary and details pages.

Change-Id: I99e272f41fea506792b1006e63b7d13e10334526

Changes for issue  "format multi select virtual network text in physical routers"

and also addressed below issues

1)IP fabric subnets object in global config should be send as null if all
removed
2)Add uuid to physical router in grid expand details
3)Have lable name as Parent Physical Inteface for Physical inf connection
4)Source Port in BGP Routers should not set in api server as 0  if not entered any input value

Closes-bug: #1545994

Change-Id: Iaa38150c807aabd644860e220b0c743cd468f754