Edit window is sometimes empty - Fixed
Issue:
The function was declear as a globle scope.
When a function is decleared without a var, it is considered as global scope.
because of that the function from one page is overwriting the function in
another page.

Fix:
Made the function with the declaration as var,
so that it will be turning to a local function and wont overwright the current
page function.

Closes-Bug: #1549806

Please enter the commit message for your changes. Lines starting

Change-Id: I8f5e452ab6eb93848ac5135855e23d0839e1f7b2