Shared Network was not getting selected in Port edit.
The shared network was not geting selected in the dropdown of the network in
Port page.
When we do a edit of a port with shared network it was getting edited with the
shared network not getting selected but the network was disabled.

and also fixed  "Port creation fails with shared network" issue

Reason:
The function formatCurrentFQName was modifing the value passed to the function
because of that the variable in the parent function is modified.

Fixed:
Made a clone of the function so the the parent function values are safe and not
changed.

Change-Id: Id4a66b03c65a5b2b8b9cf5ed56d63bc504c45afe
Closes-Bug: #1550385