Unable to add Network Policy rule with Specific IP Protocol- Fixed it
In UI Added the support for adding a port value for the protocol in policy rule and SG.
The value range from 0 to 255.
The drop down for the protocol is changed to combobox so that user defined
values can be added.

Change-Id: I51a5d5f3ac4e29eb24b8f2e5201573e70d97afc5
Closes-Bug: #1489430
(cherry picked from commit f9c79fc6af137e3ff94970e2d782a6a727d42439)