Allow Address Pair MAC address attribute itself is not added when the value is empty.
In Allow address Pair the value for Allow address pair mac is sent as empty string from UI.
Now it is changed. The attribuet itself is not getting sent if the value of the allow
address pair mac is not set.

Change-Id: Iac25d902b6752bb0b849bce72480aaf6728c5352
Partial-Bug: #1549339
(cherry picked from commit d659732ec835c5dde84aa2416735a1eda2079471)