Changes for allowing same vmi association for multiple logical interfaces
Closes-bug: #1606440

Following Manual Test cases are executed successfully

1)Create logical interface li1 with one port vmi1
2)Create logical interface li2 and map vmi1 and it should not remove vmi1 from
li1
3)Delete li1, it should not do implicit delete of vmi1 as it is referred by li2

Pending:
Now while editing port part for li we are giving option to clearport(s), this
option will be saved in cookie value, if it is true,  we will remove the port from the system.

same behavior should be extended to the grid while deleting li.

we will raise separate review for this.

Change-Id: I83b06d2e2022f92a1e7ee6bd115a3921b3148e83