Security Policy Web UI Controller changes
It is a squash commit contains below changes:
Removed Service group port validation and setting default port value.
Added implicitDeny css class for link.
Renamed tags and security sub menus as for review comments.
Replacing zero in bytes and out bytes with 1 bytes such that the data
will bind to the link element which is used to display the sessions data
in monitoring page side bar.
In case of any any in endpoint1 we are displaying the true in sidebar
changed it to any any.
We are seeing '-' instead of deny when rule action is deny, fixed it.

Related-Bug: #1695574
Change-Id: If4d9fadfda911909fc9b4789e1b3d491090f267d