Network monitoring page VN's and VMI's count mismatch is different pages
VN's and VMI's count mismatch is there between the VN & VMI
summary page compared with dashboard.

Reason is we are displaying union of config and analytics
response in VMI summary page and dashboard it is only analytics
response. In case of VN service VNs are excluded in dashboard
and included in VN summary page, fixed both issues now the
count is matching across all the pages.
Related-Bug:#1703593
Change-Id: Idd99321b01e1c2132f1ecb7c16e3f1714755ed1c