Removing the restrictions over protocol, ports and direction when Mirroring or Services is enabled on Policy rule.
There were some restrictions added in UI earlier like restrict the protocol and ports
to Any when Mirroring is enabled and Action to be Pass and Direction to be bi-directional
when Services was enabled.
As per discussions we are removing these restrictions and allowing user to enter other than Any.

Change-Id: Ia3bd8bafb83f39089732387b5410b8aace861b15
Closes-Bug: #1763519