Fixed Logical interface creation failure when there is no 'default' SG - Kubernetes
While creation of Logical Interface we were creating VMI's and attaching 'default'
SG to it. But in the case of Kubernetes the default SG name used to be k8s-<project-name>-default-sg.
To avoid the failure we are now attaching SG only if 'default' SG is present else no SG
will be attached.

Change-Id: I7d0138e4c35a32f9133cd05fae04843fc4dae43f
Closes-Bug: #1770241