Fixing the firewall-policy where adding a rule in between was getting deleted
When we add a new firewall rule in between 2 rules the sequence number
changes for all and the jsondiff was considering the modifications
as add and delete inturn causing the rule to be disassociated from
the policy.
Added the fire-rule in the exception list so that the diff ignores this.

Change-Id: I6c4f2f1863b4670185a4ba8fa88b682fd11fdbe7
Closes-Bug: #1767112